backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
starlette==0.27.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
cryptography==42.0.8

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Email Services
sendgrid==6.10.0
boto3==1.34.0  # For AWS SES
resend==0.6.0

# Push Notification Services
firebase-admin==6.2.0
pyfcm==1.5.4
pywebpush==1.14.0

# Analytics Services
google-analytics-data==0.17.4
mixpanel==4.10.0
segment-analytics-python==2.2.3

# Internationalization
babel==2.13.1
python-i18n==0.3.9

# Date/Time
python-dateutil==2.8.2
pytz==2023.3
pendulum==2.1.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Caching
redis==5.0.1
python-redis-lock==4.0.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# File Storage
boto3==1.34.0  # For S3
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0

# Image Processing
Pillow==10.1.0
python-magic==0.4.27

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Rate Limiting
slowapi==0.1.9
limits==3.6.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Currency & Finance
forex-python==1.8
CurrencyConverter==0.17.6

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Excel/CSV
openpyxl==3.1.2
xlsxwriter==3.1.9

# WebSocket Support
websockets==12.0

# Background Scheduler
apscheduler==3.10.4

# Data Export/Import
jsonlines==4.0.0
pyyaml==6.0.1

# Performance & Optimization
orjson==3.9.10
aiocache==0.12.2

# Compliance & Privacy
python-gnupg==0.5.1

# Health Checks
aioredis==2.0.1

# Production Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# New dependencies from the code block
python-decouple==3.8
# sqlite3 is built-in to Python
email-validator==2.1.0
stripe==7.8.0
groq==0.28.0
chromadb==0.4.18
sentence-transformers==2.2.2
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
pdfplumber==0.9.0
python-docx==0.8.11
aiofiles==23.2.0
tenacity==8.2.3
langchain==0.1.0
langchain-groq==0.0.3
tiktoken==0.5.2
openai==1.6.1

# Market Intelligence & Web Scraping
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
selenium==4.16.0
playwright==1.40.0
scrapy==2.11.0

# Enhanced Data Processing
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# Additional Redis features
redis-py-cluster==2.1.3
hiredis==2.2.3

# Document processing dependencies
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pdf2image==1.16.3
aiofiles==23.2.1

# Optional dependencies for better processing
poppler-utils  # For pdf2image (system dependency)

# CORS
fastapi-cors==0.0.6

# AI/ML Dependencies
openai==1.3.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Optional email service providers (install as needed)
# sendgrid==6.10.0
# boto3==1.34.0
# resend==0.6.0

# Environment management
python-dotenv==1.0.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1 